+++ src/commands/cd.js
@@ -5,7 +5,7 @@ const _errorInterpret = require("../functions/errorInt");
 const _fatalError = require("../functions/fatalError");
 const _verbInt = require("../functions/verboseInt");
 
-const verbMsgs = [
+const _verbMsgs = [
   `Replacing '/s' with spaces...`,
   `Checking if directory exists (not possible if running with '--verbose')...`,
   `Check complete, continuing...`,
@@ -21,37 +21,37 @@ const cd = (dir, ...params) => {
   let verbose = false;
   if (params.includes("--verbose") || params.includes("/verbose")) verbose = true;
 
-  _verbInt(verbMsgs[0], verbose);
+  _verbInt(_verbMsgs[0], verbose);
   dir = _replaceSpaces(dir);
 
-  _verbInt(verbMsgs[1], verbose);
+  _verbInt(_verbMsgs[1], verbose);
   if (typeof dir === "undefined") {
     _errorInterpret(2);
     return;
   }
-  _verbInt(verbMsgs[2], verbose);
+  _verbInt(_verbMsgs[2], verbose);
 
   try {
-    _verbInt(verbMsgs[3], verbose, { variable: dir });
+    _verbInt(_verbMsgs[3], verbose, { variable: dir });
     process.chdir(dir);
   } catch (err) {
     if (err.code === "ENOENT") {
-      _verbInt(verbMsgs[4], verbose, { variable: dir, error: 3 });
+      _verbInt(_verbMsgs[4], verbose, { variable: dir, error: 3 });
       _errorInterpret(3, { variable: dir });
     } else if (err.code === "EPERM") {
-      _verbInt(verbMsgs[5], verbose, { variable: dir, error: 4 });
+      _verbInt(_verbMsgs[5], verbose, { variable: dir, error: 4 });
       _errorInterpret(4, { variable: dir });
     } else {
-      _verbInt(verbMsgs[6], verbose, { variable: dir, error: err.code });
+      _verbInt(_verbMsgs[6], verbose, { variable: dir, error: err.code });
       _fatalError(err);
     }
 
-    _verbInt(verbMsgs[7], verbose, { variable: dir, error: err.code });
+    _verbInt(_verbMsgs[7], verbose, { variable: dir, error: err.code });
     return;
   }
 
   console.log(`Changed directory to ${chalk.green(process.cwd())}.\n`);
-  _verbInt(verbMsgs[8], verbose, { variable: dir });
+  _verbInt(_verbMsgs[8], verbose, { variable: dir });
 };
 
 module.exports = cd;