+++ src/commands/copyfile.js
@@ -8,19 +8,49 @@ const _convertAbsolute = require("../functions/convAbs");
 
 const _errorInterpret = require("../functions/errorInt");
 const _fatalError = require("../functions/fatalError");
+const _verbInt = require("../functions/verboseInt");
 
-const copyfile = (src, dest) => {
+const _verbMsgs = [
+  `Replacing spaces in 'src' and 'dest'...`,
+  `Checking if 'src'/'dest' is not provided...`,
+  `Either 'src' or 'dest' is not provided (is 'undefined') (error code %ERROR%)!`,
+  `There was an error, and the 'copyfile' command has been exited.`,
+  `Removing starting and ending whitespace from 'src' and 'dest'...`,
+  `Prompting for confirmation...`,
+  `Checking if the operation was accepted or declined...`,
+  `The operation was cancelled (error code %ERROR%)!`,
+  `Operation accepted, continuing...`,
+];
+
+const copyfile = (src, dest, ...params) => {
+  let verbose = false;
+  if (
+    params.includes("--verbose") ||
+    params.includes("/verbose") ||
+    src === "--verbose" ||
+    src === "/verbose" ||
+    dest === "--verbose" ||
+    dest === "/verbose"
+  )
+    verbose = true;
+
+  _verbInt(_verbMsgs[0], verbose);
   src = _replaceSpaces(src);
   dest = _replaceSpaces(dest);
 
+  _verbInt(_verbMsgs[1], verbose);
   if (!src || !dest) {
+    _verbInt(_verbMsgs[2], verbose, { error: 5 });
     _errorInterpret(5);
+    _verbInt(_verbMsgs[3], verbose);
     return;
   }
 
+  _verbInt(_verbMsgs[4], verbose);
   const srcPath = _convertAbsolute(src.trim());
   const destPath = _convertAbsolute(dest.trim());
 
+  _verbInt(_verbMsgs[5], verbose);
   const confirmText = readline
     .question(
       `${chalk.bold.red("WARNING!")} If ${chalk.bold(
@@ -30,10 +60,14 @@ const copyfile = (src, dest) => {
       )}/${chalk.red.bold("N")}] `
     )
     .toLowerCase();
+  _verbInt(_verbMsgs[6], verbose);
   if (confirmText.includes("n") || !confirmText.includes("y")) {
+    _verbInt(_verbMsgs[7], verbose, { error: 6 });
     _errorInterpret(6);
+    _verbInt(_verbMsgs[3], verbose);
     return;
   }
+  _verbInt(_verbMsgs[8], verbose);
 
   try {
     console.log(chalk.italic.blueBright("Please wait...\n"));
@@ -50,6 +84,7 @@ const copyfile = (src, dest) => {
     } else {
       _fatalError(err);
     }
+    return;
   }
 
   console.log(chalk.green("The operation completed successfully.\n"));